Refresh centerline csv's instead of appending alias'
